Fix warnings

+++ input/input_autodetect.c
@@ -232,7 +232,7 @@ static bool input_autoconfigure_joypad_from_conf_dir(
    else
    {
       RARCH_LOG("Autodetect: no profiles found for %s (%d/%d)", params->name, params->vid, params->pid);
-      snprintf(msg, sizeof(msg), "%s (%ld/%ld) not configured", params->name, params->vid, params->pid);
+      snprintf(msg, sizeof(msg), "%s (%ld/%ld) not configured", params->name, (long)params->vid, (long)params->pid);
       rarch_main_msg_queue_push(msg, 0, 60, false);
       ret = 0;
    }
